skills/pandarey0130-lab/community-os-v2/requirements.txt
python-telegram-bot>=21.0
chromadb>=0.4.0
pymupdf>=1.23.0
python-docx>=1.0.0
pyyaml>=6.0
requests>=2.31.0
apscheduler>=6.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
# ── 知识库 & RAG ─────────────────────────────────────
langchain>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
sentence-transformers>=2.2.0
watchdog>=3.0.0
hnswlib>=0.7.0
sseclient-py
python-multipart>=0.0.9
